POS/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-admin-rangefilter==0.8.2
django-crum==0.7.9
django-debug-toolbar==3.2.2
django-easy-pdf==0.1.1
django-import-export==2.6.1
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
PyYAML==6.0
reportlab==3.6.2
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
